backend/requirements.txt
Flask==2.2.3
flask_cors==3.0.10
PyJWT==2.6.0
Werkzeug==2.2.3
pytest==7.2.2
requests==2.28.2
openai==0.27.4
pymongo==4.3.3
